+++ src/lib/onedata-gui-websocket-client/tests/dummy/app/session-stores/application.js
@@ -1,0 +1,1 @@
+export { default } from 'onedata-gui-websocket-client/session-stores/application';