+++ CHANGELOG.md
@@ -6,5 +6,6 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-3676 Support for Graph API push messages (update/delete)
 * VFS-3883 Porting recent improvements in common components and utils from op-gui-default
 * VFS-3547 Prototype of Providers view (using backend mock)